refactored merge files
