+++ src/custom-recognition/merges/version06.js
@@ -1,0 +1,9 @@
+async function mergveVersion06(data) {
+
+    const meleeObject = data.melee;
+
+    const newMeleeObject = {
+        id: 
+    }
+    
+}
