+++ src/custom-recognition/merges/versionIndex.js
@@ -1,0 +1,5 @@
+export {mergeVersion01} from "./version01.js"
+export {mergeVersion02} from "./version02.js"
+export {mergeVersion03} from "./version03.js"
+export {mergeVersion04} from "./version04.js"
+export {mergeVersion05} from "./version05.js"
